move representation calculation to a separate module
